// src/interfaces/bo.ts
export interface ICaracteristicas {
  ID_ENV_CARACTERISTICA: string
  ALTURA_APARENTE: string
  PESO: string
  IDADE_APARENTE: string
  ID_COR_CABELO: string
  NM_COR_CABELO: string
  ID_TIPO_PELAGEM_FACIAL: string
  NM_TIPO_PELAGEM_FACIAL: string
  ID_TIPO_APARENCIA: string
  NM_TIPO_APARENCIA: string
  ID_TIPO_CABELO: string
  NM_TIPO_CABELO: string
  ID_COR_PELE: string
  NM_COR_PELE: string
  ID_COR_OLHOS: string
  NM_COR_OLHOS: string
  ID_TIPO_MARCA_FISICA: string
  NM_TIPO_MARCA_FISICA: string
  TATUAGEM: string
  TIPO_DENTES: string
  TIPO_CICATRIZ: string
  BIGODE: string
  ID_TIPO_DEFEITO_FISICO: string
  NM_TIPO_DEFEITO_FISICO: string
  ID_TIPO_DEFICIENCIA: string
  NM_TIPO_DEFICIENCIA: string
}